+++ lib/azure/storage/core/auth/shared_access_signature.rb
@@ -25,4 +25,4 @@
 require 'azure/storage/core/auth/shared_access_signature_generator'
 require 'azure/storage/core/auth/shared_access_signature_signer'
 
-include Azure::Storage::Auth
+include Azure::Storage::Core::Auth